move external event to it's own file
